fix(dev): slow types

+++ packages/mikro-orm/src/bsqlite/index.ts
@@ -1,5 +1,5 @@
 import { BetterSqliteDriver } from '@mikro-orm/better-sqlite'
-import { MikroORM, Options } from '@mikro-orm/core'
+import { MikroORM, type Options } from '@mikro-orm/core'
 import { TsMorphMetadataProvider } from '@mikro-orm/reflection'
 import { SqlHighlighter } from '@mikro-orm/sql-highlighter'
 import { defaultsDeep } from '@vegapunk/utilities'
@@ -17,16 +17,18 @@ const baseOptions = {
 	allowGlobalContext: true,
 } satisfies BSqliteOptions & { driver: unknown }
 
-export async function start(options: BSqliteOptions): Promise<MikroORM> {
+export async function start(options: BSqliteOptions, autoSchema?: boolean): Promise<MikroORM> {
 	const db = await MikroORM.init(options)
 
-	const generator = db.getSchemaGenerator()
-	if (process.env.NODE_ENV === 'development') {
-		await generator.dropSchema()
-		await generator.createSchema()
-	} else {
-		await generator.ensureDatabase()
-		await generator.updateSchema()
+	if (autoSchema) {
+		const generator = db.getSchemaGenerator()
+		if (process.env.NODE_ENV === 'development') {
+			await generator.dropSchema()
+			await generator.createSchema()
+		} else {
+			await generator.ensureDatabase()
+			await generator.updateSchema()
+		}
 	}
 	return db
 }
